IPBL/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Mail==0.9.1
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0
cryptography==41.0.7
gunicorn==21.2.0
psycopg[binary]

